Start vscode in current directory by default
Current behavior is to start in $REPO_DIR if $CODE_WORKINGDIR is
not set. However, in JupyterHubs, you want to default to starting in
current working directory, which is often set to /home/jovyan or
similar. Current working directory is already set to $REPO_DIR in
mybinder.org by default, so this should be a no-op in mybinder.org
while doing the expected thing on JupyterHubs.